refactor: split recipe schema pattern validation
